chore(v3/organizations): cleanup and reorganize
